Fix uint64 & int64 typedef

+++ include/types.h
@@ -3,9 +3,9 @@
 typedef unsigned char uint8;
 typedef unsigned short uint16;
 typedef unsigned int uint32;
-typedef unsigned long uint64;
+typedef unsigned long long uint64;
 
 typedef char int8;
 typedef short int16;
 typedef int int32;
-typedef long int64;
+typedef long long int64;
